fix: change method of shutting down mitmdump to fix ctrl+c issue in unsetting proxy env vars in windows nuitka build

+++ pomodoro_task_app/website_blocker/utils.py
@@ -9,6 +9,7 @@ import signal
 import subprocess
 
 import psutil
+from loguru import logger
 
 
 def exec_command(command):
@@ -32,17 +33,29 @@ def find_processes_by_name(name):
 
 
 def kill_process():
+    logger.debug("Inside kill_process().")
     if os.name == "nt":
         processes = find_processes_by_name("mitmdump.exe") + find_processes_by_name("mitmproxy.exe")
         for p in processes:
-            try:
-                p.send_signal(signal.CTRL_C_EVENT)
-            except (psutil.AccessDenied, AttributeError):
-                p.kill()
+            logger.debug(f"Trying to kill process with pid {p.pid} on Windows.")
+            p.kill()  # I couldn't find any way of stopping mitmdump gracefully on Windows
     else:
+
         processes = find_processes_by_name("mitmdump") + find_processes_by_name("mitmproxy")
+        # list all processes with id and name
+        for p in processes:
+            logger.debug(f"Found process with pid {p.pid} and name {p.info['name']}.")
+        # if no process is found then
+        if not processes:
+            logger.debug("No mitmdump/mitmproxy process found.")
+            return
+
         for p in processes:
             try:
+                logger.debug(f"Trying to gracefully shutdown mitmdump with pid {p.pid} on Linux/MacOS "
+                             f"inside kill_process().")
                 p.send_signal(signal.SIGINT)
             except psutil.AccessDenied:
+                logger.error(f"Access denied to kill process with pid {p.pid}.")
+                logger.debug(f"Trying to kill process with pid {p.pid}.")
                 p.kill()
